fix form2, se modularizo
