+++ client/src/utils/formInputs.js
@@ -1,0 +1,18 @@
+export const inputNumber = [
+  {value: "floors", inputName: "Pisos"},
+  {value: "enviroments", inputName: "Ambientes"},
+  {value: "rooms", inputName: "Habitaciones"},
+  {value: "bathrooms", inputName: "Baños"},
+  {value: "garage", inputName: "Cochera"},
+  {value: "area", inputName: "Area de la propiedad en M²"},
+  {value: "antiquity", inputName: "Antiguedad en años"}
+];
+export const inputServices = [
+  {inputName: "Luz", value: "ligth"},
+  {inputName: "Agua", value: "water"},
+  {inputName: "Gas", value: "gas"},
+  {inputName: "Cloaca", value: "sewer"},
+  {inputName: "Wifi", value: "Wifi"},
+  {inputName: "Cable", value: "tv"},
+  {inputName: "Telefono", value: "telephone"},
+]
